requirements.txt
# 웹 크롤링 관련
selenium>=4.15.0
webdriver-manager>=4.0.0

# 시스템 모니터링
psutil>=5.9.0

# 데이터 처리 및 저장
pandas>=2.0.0
requests>=2.31.0

# CLI 및 유틸리티
tqdm>=4.65.0
click>=8.1.0

# 로깅 및 설정
python-dotenv>=1.0.0

# 데이터 검증
jsonschema>=4.17.0

# 날짜/시간 처리
python-dateutil>=2.8.0

# 텍스트 처리
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 구글 드라이브 / 서비스계정
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0